% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{block_matrix}
\alias{block_matrix}
\alias{minus_block_matrix}
\title{Block matrix}
\usage{
block_matrix(x, s)

minus_block_matrix(x, s)
}
\arguments{
\item{x}{a square matrix}

\item{s}{the dimensions (width = height) of the block matrix being selected over the diagonal of matrix x}

\item{x}{a square matrix}

\item{s}{the dimensions (width = height) of the block matrix being dropped over the diagonal of matrix x}
}
\description{
Block matrix

Minus Block matrix
}

