% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfddva.R
\name{dfddva}
\alias{dfddva}
\title{Domestic Final Demand Domestic Value Added}
\usage{
dfddva(x, aggregate = FALSE)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr, which should be post multiplied with final demand (using the parameter: post="final_demand")}

\item{aggregate}{should dfddva be aggregated along source industries to a national sum?}
}
\description{
Domestic Final Demand Domestic Value Added
}
\examples{
# load the decompr package
library(decompr)

# load example data
data(leather)

# create a leontief decomposed data set
l <- decomp(inter,
            final,
            countries,
            industries,
            out,
            method = "leontief",
            post = "final_demand")
 
 # apply dfddva
 dfddva( l )
}

