% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffddva.R
\name{ffddva}
\alias{ffddva}
\title{Foreign Final Demand Domestic Value Added}
\usage{
ffddva(x, aggregate = FALSE)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr, which should be post multiplied with final demand (using the parameter: post="final_demand")}

\item{aggregate}{should dfddva be aggregated along source industries to a national sum?}
}
\description{
Foreign Final Demand Domestic Value Added
}
\examples{
# load the decompr package
library(decompr)

# load example data
data(leather)
attach(leather)

# create a leontief decomposed data set
l <- decomp(x = inter,
            y = final,
            k = countries,
            i = industries,
            o = out,
            method = "leontief",
            post = "final_demand")

 # apply ffddva
 ffddva( l )
}
