% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2e.R
\name{i2e}
\alias{i2e}
\alias{vertical_specialisation}
\alias{vertical_specialization}
\title{Importing to Export}
\usage{
i2e(x, by = NULL, subset = NULL)

vertical_specialisation(x, by = NULL, subset = NULL)

vertical_specialization(x, by = NULL, subset = NULL)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr}

\item{by}{variable to subset by}

\item{subset}{value(s) of the subset variable to select}
}
\description{
Importing to Export

Vertical Specialization

Vertical Specialisation
}
\examples{
# load the decompr package
library(decompr)

# load the example data set
data(leather)
attach(leather)

# create a leontief decomposed data set
l <- decomp(x = inter,
            y = final,
            k = countries,
            i = industries,
            o = out)
 
 # apply the Import to Exports analysis
 i2e( l )
}
