% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minus_block_matrix.R
\name{minus_block_matrix}
\alias{minus_block_matrix}
\title{Minus Block matrix}
\usage{
minus_block_matrix(x, s)
}
\arguments{
\item{x}{a square matrix}

\item{s}{the dimensions (width = height) of the block matrix being dropped over the diagonal of matrix x}
}
\description{
Minus Block matrix
}

