% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rca.R
\name{rca}
\alias{rca}
\title{Revealed Comparative Advantage}
\usage{
rca(data, country, sector)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr}
}
\description{
Revealed Comparative Advantage
}

