% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/vertical_specialisation.R
\name{vertical_specialisation}
\alias{vertical_specialisation}
\title{Vertical Specialisation}
\usage{
vertical_specialisation(x)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr}
}
\description{
Vertical Specialisation
}
\examples{
# load the decompr package
library(decompr)

# load the example data set
data(leather)

# create a leontief decomposed data set
l <- decomp(inter,
            final,
            countries,
            industries,
            out, long=FALSE)

 # apply the vertical_specialisation analysis
 vertical_specialisation(l)
}

