% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{block_matrix}
\alias{block_matrix}
\title{Block matrix}
\usage{
block_matrix(x, s)
}
\arguments{
\item{x}{a square matrix}

\item{s}{the dimensions (width = height) of the block matrix being selected over the diagonal of matrix x}
}
\description{
Block matrix
}

