% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfddva.R
\name{dfddva}
\alias{dfddva}
\title{Domestic Final Demand Domestic Value Added}
\usage{
dfddva(x)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr, which should be post multiplied with final demand (using the parameter: post="final_demand")}
}
\description{
Domestic Final Demand Domestic Value Added
}
\examples{
# load the decompr package
library(decompr)

# load the example data set
data(leather)

# create a leontief decomposed data set
l <- decomp(inter,
            final,
            countries,
            industries,
            out,
            method = "leontief",
            post = "final_demand")

 # apply dfddva
 dfddva( l )
}

