% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/e2r.R
\name{e2r}
\alias{e2r}
\title{Exporting to Re-export}
\usage{
e2r(x)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr}
}
\description{
Exporting to Re-export
}
\examples{
# load the decompr package
library(decompr)

# load the example data set
data(leather)

# create a leontief decomposed data set
l <- decomp(inter,
            final,
            countries,
            industries,
            out)

 # apply the Exporting to Re-export
 e2r( l )
}

