% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/i2e.R
\name{i2e}
\alias{i2e}
\alias{vertical_specialisation}
\alias{vertical_specialization}
\title{Importing to Export}
\usage{
i2e(x)

vertical_specialisation(x)

vertical_specialization(x)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr}
}
\description{
Importing to Export

Vertical Specialization

Vertical Specialisation
}

