% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/nrca.R
\name{nrca}
\alias{nrca}
\title{New Revealed Comparative Advantage}
\usage{
nrca(data, country, sector)
}
\arguments{
\item{data}{A Wang-Wei-Zhu decomposed Inter-Country Input Output table as created by decompr}

\item{country}{target country}

\item{sector}{target sector}
}
\description{
New Revealed Comparative Advantage
}

