% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rca.R
\name{rca}
\alias{rca}
\title{Revealed Comparative Advantage}
\usage{
rca(data, country, sector)
}
\arguments{
\item{data}{A Leontief decomposed Inter-Country Input Output table as created by decompr}

\item{country}{target country}

\item{sector}{target sector}
}
\description{
Revealed Comparative Advantage
}

