% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upstream.R
\name{upstream}
\alias{upstream}
\title{Upstreamness}
\usage{
upstream(x)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr, which should be post multiplied with final demand (using the parameter: post="final_demand")}
}
\description{
Upstreamness
}
\examples{
# load the decompr package
library(decompr)

# load example data
data(leather)

# create a leontief decomposed data set
l <- load_tables_vectors(inter,
            final,
            countries,
            industries,
            out)

 # apply upstream
 upstream( l )
}

