% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vertical_specialisation.R
\name{vertical_specialisation}
\alias{vertical_specialisation}
\title{Vertical Specialisation}
\usage{
vertical_specialisation(x)
}
\arguments{
\item{x}{A Leontief decomposed Inter-Country Input Output table as created by decompr}
}
\description{
Vertical Specialisation
}

